// front/src/types/user.ts
export interface userType {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  profilePic: string;
  mobile: string;
  document: string;
  zipCode: string;
  street: string;
  number: string;
  complement: string;
  neighborhood: string;
  city: string;
  state: string;
  isClient: boolean;
  isCollector: boolean;
  createdAt: Date;
}